New features added. Replaying function, as well as, another video function now works properly. With the addition of a guide on player control. Such as how to use each feature, etc, general information about how-to.
Added a quit function, that will quit out of the music player. Effectively closing it. To reopen it, do what you did the first time.

However, a new issue is found. When selecting the first video, it will always choose the last video. Reasons for this is unknown as the video selection feature has been tested multiple times. A new file will be commited within a few hours, detailing the issue. Any help will be appreciated.