Files added to .gitignore, and changes regarding flow of control was made in script.py
.gitignore
Added a new folder to untrack, GUI, as this is not the GUI branch there was no reason to track it.

script.py
There was an issue with the flow of control, as mentioned above. When using one of the commands from PlayerControl, it kept asking the using if they wanted to watch another video. Or it will just choose to not, ask the user, and display the previous search results. This has been solved.

A few methods, namely; SkipToNextVideo, and NextVideos will be incredibly difficult to implement without a GUI. Any help will be greatly appreciated. If you wish to contribute, go ahead and find my discord tag in the README.md file, and dm me.