Added files to .gitignore, GUI.py is not able to work with edit.py
GUI.py
The GUI component of UMA. GUI.py is now able to interact with edit.py. However, only the replay button works properly. Download, and Play do not. As it only pauses for a moment. Then it will play again. Download can't work, due to not being able to request if the user want's to download an mp3 file or an mp4, since that input cannot be completed it will not work.

edit.py, renamed from copy.py
The reason for the name change was because copy was an import in the PyTube module, and there were conflicts. Changing the name to edit fixed that.

edit.py has been drastically changed, with many new functions added. The most important one is ReadFromCommand(). In GUI.py there is a command called WriteToCommand(). These are what I called 'bridging' functions. Commands.txt is the bridge. That file will be created, and commands will be written to, and commands will be taken from there as well. It acts as a bridge. Hence the name 'bridging' or 'bridge' functions.